python-requirements.txt
lxml==4.4.2
odfpy==1.4.0
